Post-Thursday fixes
first off, we aren't using `sessionStorage` anymore. wasn't providing any speed
increases, and we could get the information other ways. this also means that
the pages no longer require the index to be loaded first. Secondly, the page
names ('Chapter' names) are static now, so each page can find itself in the
sequence. so it's basically where it was a week ago. Now, we expand it for
sections and figures.